listo el scheduler completamente funcional, junto con el mail, agregados tambien los reportes de retraso en caso de falla del servidor con notificaciones pasadas a 5 minutos

+++ proyecto/config/initializers/task_scheduler.rb
@@ -1,14 +1,22 @@
 require 'rubygems'
-require 'rufus/scheduler' 
+require 'rufus/scheduler'
+require 'chronic'
 
 scheduler = Rufus::Scheduler.start_new
 
 scheduler.every '5m' do
-	@note = Note.all
 	Note.all.each do |n|
-		if (Time.parse(n.delivery.to_s) - Time.now <= 300)
-			ReminderMail.new_remind_me_email(n).deliver
-			n.destroy
+		timex = Time.now - Chronic.parse(n.delivery.to_s)
+		puts timex
+		if (timex >= 0)
+			if (timex <= 300)
+				ReminderMail.new_remind_me_email(n).deliver
+				n.destroy
+			else
+				n.content += " [Sorry for the delay, seems like we had problems!]"
+				ReminderMail.new_remind_me_email(n).deliver
+				n.destroy
+			end
 		end
 	end
 end
